UPDATE: Move TensorflowTrainer into Backend
