+++ use-cases/cyclones/trainer.py
@@ -1,59 +1,14 @@
 import tensorflow.keras as keras
-import logging
 import tensorflow as tf
 
-from datetime import datetime
-from tqdm import tqdm
-from os.path import join, exists
-from lib.strategy import get_mirrored_strategy
-from lib.models.setup import get_network_config, load_model
+from os.path import join
+from lib.models.setup import get_network_config
 from lib.callbacks import ProcessBenchmark
 from lib.macros import (
-    PatchType,
-    Network,
     Losses,
     RegularizationStrength,
-    Activation,
-    LabelNoCyclone,
-    AugmentationType,
 )
-
-# TODO: Abstraction for Trainer
-class TensorflowTrainer:
-    def __init__(self, strategy, loss, epochs, batch_size, callbacks, optimizer, model_func, metrics_func):
-        self.strategy = strategy
-        self.epochs = epochs
-        self.batch_size = batch_size
-        self.loss = loss
-        self.callbacks = callbacks
-        self.global_batch_size = self.batch_size * self.strategy.num_replicas_in_sync
-        self.optimizer = optimizer
-
-        # Create distributed TF vars
-        with self.strategy.scope():
-            self.model = model_func()
-            self.model.compile(loss=self.loss, optimizer=self.optimizer, metrics=metrics_func())
-
-    def train(self, data):
-        (train, n_train), (test, n_test) = data
-        train = self.strategy.experimental_distribute_dataset(train)
-        test = self.strategy.experimental_distribute_dataset(test)
-
-        # compute the steps per epoch for train and valid
-        train_steps = n_train // self.batch_size
-        test_steps = n_test // self.batch_size
-
-        # train the model
-        self.model.fit(
-            train,
-            validation_data=test,
-            steps_per_epoch=train_steps,
-            validation_steps=test_steps,
-            epochs=self.epochs,
-            callbacks=self.callbacks,
-        )
-
-        logging.debug(f"Model trained")
+from itwinai.backend.tensorflow.trainer import TensorflowTrainer
 
 class CyclonesTrainer(TensorflowTrainer):
     def __init__(
@@ -77,8 +32,6 @@ class CyclonesTrainer(TensorflowTrainer):
         optimizer = keras.optimizers.Adam(learning_rate=learning_rate)
 
         # Paths, Folders
-        SCALER_DIR = join(RUN_DIR, "scalers")
-        TENSORBOARD_DIR = join(RUN_DIR, "tensorboard")
         CHECKPOINTS_DIR = join(RUN_DIR, "checkpoints")
 
         # files and csvs definition
@@ -124,6 +77,4 @@ class CyclonesTrainer(TensorflowTrainer):
         )
 
     def train(self, data):
-        # TODO: Integrate train_step locally and pass it into the super distributed training function
         super().train(data)
-
